backend: 重构 IR -> LIR
- 进一步抽出了生成 LIR 的部分
- 抽出了做寄存器分配的部分
